requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
altair==4.1.0
ansi2html==1.5.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrdict==2.0.1
attrs==20.3.0
autoflake==1.4
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
black==19.10b0
bleach==3.2.1
blessings==1.7
blinker==1.4
blis==0.4.1
bokeh==2.2.3
boto==2.49.0
boto3==1.4.8
botocore==1.8.50
bravado==11.0.2
bravado-core==5.17.0
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.1
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.4
colorful==0.5.4
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.3
cytoolz==0.11.0
dash==1.18.1
dash-bootstrap-components==0.10.6
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
dask==2.20.0
dask-labextension==3.0.0
dataclasses==0.6
datasketch==1.5.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distributed==2.30.1
distro==1.5.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
elasticsearch==7.8.0
en-core-web-lg==2.3.1
en-core-web-sm==2.3.1
entrypoints==0.3
errordetection==0.0.1
faiss-gpu==1.6.3
fastapi==0.63.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Compress==1.8.0
Flask-HTTPAuth==4.1.0
fsspec==0.8.3
ftfy==4.4.3
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
google==3.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.32.0
gym==0.17.2
h11==0.11.0
HeapDict==1.0.1
hiredis==1.1.0
html5lib==1.1
httptools==0.1.1
hydra-core==1.0.0rc2
idna==2.10
imbalanced-learn==0.7.0
importlib-metadata==2.0.0
importlib-resources==3.0.0
ipycallback==0.2.5
ipyevents==0.8.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-dash==0.3.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.4.1
jupyter-server-proxy==1.5.0
jupyterlab==2.2.8
jupyterlab-code-formatter==1.3.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
lm-scorer==0.4.2
locket==0.2.0
loguru==0.5.1
lxml==4.5.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
modin==0.8.2
monotonic==1.5
msgpack==1.0.1
multidict==4.7.6
munkres==1.1.4
murmurhash==1.0.2
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neptune-client==0.4.130
neptune-contrib==0.25.0
nest-asyncio==1.4.3
networkx==2.3
nltk==3.5
nose==1.3.7
notebook==6.1.5
numpy==1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.1rc11
opencensus==0.7.11
opencensus-context==0.1.2
packaging==20.8
pandarallel==1.5.1
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.3.3
plac==1.1.3
plotly==4.10.0
plumbum==1.6.9
preshed==3.0.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pslpython==2.2.2
psutil==5.7.2
ptyprocess==0.6.0
py-spy==0.3.3
py3nvml==0.2.6
py4j==0.10.9
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
pyglet==1.5.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.5.3
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.0.0
pyspellchecker==0.5.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-json-logger==0.1.11
python-rapidjson==0.9.1
pytorch-lightning==1.0.8
pytorch-metric-learning==0.9.94
pytorch-nlp==0.5.0
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
ray==1.0.1.post1
redis==3.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3987==1.3.8
rpyc==4.1.5
rsa==4.6
s3transfer==0.1.13
sacremoses==0.0.43
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
sentence-transformers==0.3.6
sentencepiece==0.1.91
setuptools==51.0.0.post20201207
simpervisor==0.3
simplejson==3.17.2
six==1.15.0
smart-open==2.1.0
smmap==3.0.4
snakeviz==2.1.0
snorkel==0.9.6
sortedcontainers==2.3.0
soupsieve==2.1
spacy==2.3.2
SQLAlchemy==1.3.21
srsly==1.0.2
starlette==0.13.6
strict-rfc3339==0.7
strsimpy==0.1.9
swagger-spec-validator==2.7.3
swifter==1.0.7
tblib==1.7.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.4.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
texttable==1.6.2
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.1
toolz==0.11.1
torch==1.7.0
torchaudio==0.6.0a0+f17ae39
torchtext==0.8.0
torchvision==0.7.0
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
transformers==3.1.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.0.0
urllib3==1.26.2
usaddress==0.5.10
uvicorn==0.13.2
uvloop==0.14.0
vega-datasets==0.8.0
wasabi==0.7.1
watchgod==0.6
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.11.2
xgboost==1.3.1
XlsxWriter==1.3.3
xmltodict==0.12.0
yarl==1.5.1
zict==2.0.0
zipp==3.4.0
